+++ consultation_analyser/processing.py
@@ -1,13 +1,13 @@
 from django.conf import settings
 
 from consultation_analyser.batch_calls import BatchJobHandler
-from consultation_analyser.consultations.llm_summariser import create_llm_summaries_for_consultation
 from consultation_analyser.hosting_environment import HostingEnvironment
+from consultation_analyser.pipeline.llm_summariser import create_llm_summaries_for_consultation
 
 
 def process_consultation_themes(consultation):
     # Import only when needed
-    from consultation_analyser.consultations.ml_pipeline import save_themes_for_consultation
+    from consultation_analyser.pipeline.ml_pipeline import save_themes_for_consultation
 
     save_themes_for_consultation(consultation.id)
     create_llm_summaries_for_consultation(consultation)
