+++ tests/integration/test_ml_pipeline.py
@@ -2,7 +2,7 @@ import pytest
 
 from consultation_analyser import factories
 from consultation_analyser.consultations import models
-from consultation_analyser.consultations.ml_pipeline import (
+from consultation_analyser.pipeline.ml_pipeline import (
     save_themes_for_consultation,
 )
